kyc-plugin-django/requirements.txt
asn1crypto==0.23.0
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
cryptography==2.0.3
Django==1.11.6
gunicorn==19.7.1
idna==2.6
pycparser==2.18
PyJWT==1.5.3
pytz==2017.2
requests==2.18.4
six==1.11.0
urllib3==1.22
whitenoise==3.3.1
